requirements.txt
winocr
pillow
pdf2image
pyinstaller
numpy
